Script to parse and trim alignment based on xlsx
Draft script to parse alignment in fasta format, using Bio python AlignIO and SeqIO. Takes a fasta and XLSX input file, with an index specifying a Range of columns (sites) to delete from a fasta alignment, separated by a '-', leaving rows (taxa) unchanged. This script needs more modification before it can be used with another alignment and as a standalone python script that can be run in a shell.